fix: fix lint-staged scripts
